Updated the GUI to use ScreenManager and got createAccount to work with the new GUI
